220071601270_Decentralized_Voting_System/requirements.txt
Flask==2.3.3
web3==6.0.0
py-solc-x==1.1.1
eth-account==0.8.0